@@ -2,6 +2,8 @@ class AdminController < ApplicationController
 
   before_filter :authenticate_user!
   before_filter :admin_user?
+#  before_filter :validate_authorization!
+#  add_breadcrumb 'admin', :admin_users_path
 
   def users
     @users = User.all
@@ -15,11 +17,16 @@ class AdminController < ApplicationController
     toggle_admin_privilidges(params[:username], :demote)
   end
 
-  def destroy
-    user = User.first(:conditions => {:username => params[:username]})
+  def disable
+    user = User.find_by_username(params[:username]);
+    disabled = params[:disabled].to_i == 1
     if user
-      user.destroy
-      render :text => "removed"
+      user.update_attribute(:disabled, disabled)
+      if (disabled)
+        render :text => "<a href=\"#\" class=\"disable\" data-username=\"#{user.username}\">disabled</span>"
+      else
+        render :text => "<a href=\"#\" class=\"enable\" data-username=\"#{user.username}\">enabled</span>"
+      end
     else
       render :text => "User not found"
     end
@@ -38,14 +45,15 @@ class AdminController < ApplicationController
   private
 
   def toggle_admin_privilidges(username, direction)
-    user = User.first(:conditions => {:username => username})
+    user = User.find_by_username username
+
     if user
       if direction == :promote
         user.update_attribute(:admin, true)
-        render :text => "Yes - <span class=\"demote\" data-username=\"#{username}\">remove admin rights</span>"
+        render :text => "Yes - <a href=\"#\" class=\"demote\" data-username=\"#{username}\">remove admin rights</a>"
       else
         user.update_attribute(:admin, false)
-        render :text => "No - <span class=\"promote\" data-username=\"#{username}\">grant admin rights</span>"
+        render :text => "No - <a href=\"#\" class=\"promote\" data-username=\"#{username}\">grant admin rights</a>"
       end
     else
       render :text => "User not found"
